backend/pyproject.toml
[project]
name = "chatsune-backend"
version = "0.1.0"
description = "Chatsune backend — privacy-first AI companion platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "motor>=3.7",
    "redis>=5.2",
    "pyjwt>=2.10",
    "bcrypt>=4.2",
    "email-validator>=2.2",
    "cryptography>=46.0.6",
    "tiktoken>=0.9",
    "pillow>=11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["../tests"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
]
